[level 2] Title: 시소 짝꿍, Time: 46.77 ms, Memory: 102 MB -BaekjoonHub

+++ 프로그래머스/2/152996. 시소 짝꿍/시소 짝꿍.java
@@ -1,33 +1,33 @@
 import java.util.*;
-
+//Q: 왜 Double아니라 Long이나 Integer로 하면 틀리는거지? (어떤 경우가 있지)
 class Solution {
     private final int MAX = 1_000;
     
     public long solution(int[] weights) {
         long answer = 0;
         
-        Map<Long, Long> weightCnt = new TreeMap<>();        
+        Map<Double, Integer> weightCnt = new TreeMap<>();        
         for (int i=0;i<weights.length;i++) {
-            long cnt = weightCnt.getOrDefault((long)weights[i], 0L) + 1;
-            weightCnt.put((long)weights[i], cnt);
+            int cnt = weightCnt.getOrDefault((double)weights[i], 0) + 1;
+            weightCnt.put((double)weights[i], cnt);
         }
         
-        for (long key : weightCnt.keySet()) {
+        for (double key : weightCnt.keySet()) {
             long cnt = weightCnt.get(key);
             if (cnt > 1) {
                 answer += cnt * (cnt-1) / 2;
             }
             
-            if (weightCnt.containsKey((long)(3*key/2.0))) {
-                answer += cnt * weightCnt.get((long)(3*key/2.0));
+            if (weightCnt.containsKey(3*key/2)) {
+                answer += cnt * weightCnt.get(3*key/2);
             }
             
-            if (weightCnt.containsKey((long)(4*key/2.0))) {
-                answer += cnt * weightCnt.get((long)(4*key/2.0));
+            if (weightCnt.containsKey(4*key/2)) {
+                answer += cnt * weightCnt.get(4*key/2);
             }
             
-            if (weightCnt.containsKey((long)(4*key/3.0))) {
-                answer += cnt * weightCnt.get((long)(4*key/3.0));
+            if (weightCnt.containsKey(4*key/3)) {
+                answer += cnt * weightCnt.get(4*key/3);
             }
 
         }
